chore: move site to own repo

+++ packages/react/rollup.config.js
@@ -1,15 +1,36 @@
-import fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 import { transformSync } from 'esbuild';
+import { execa } from 'execa';
+import fs from 'fs-extra';
 import { defineConfig } from 'rollup';
 import dts from 'rollup-plugin-dts';
 import esbuildPlugin from 'rollup-plugin-esbuild';
 
-const MODE_WATCH = process.argv.includes('-w');
-const MODE_TYPES = process.argv.includes('--config-types');
-const EXTERNAL = ['react', 'react-dom', 'media-icons', 'media-captions', 'hls.js', /@radix-ui/];
-const NPM = [define({ dev: true }), define({ dev: false })];
+const __dirname = path.dirname(fileURLToPath(import.meta.url)),
+  ROOT_DIR = path.resolve(__dirname, '.'),
+  STYLES_DIR = path.resolve(ROOT_DIR, 'player/styles'),
+  VIDSTACK_PKG_DIR = path.resolve(ROOT_DIR, 'node_modules/vidstack'),
+  VIDSTACK_PKG_PLAYER_STYLES_DIR = path.resolve(VIDSTACK_PKG_DIR, 'player/styles');
+
+const MODE_WATCH = process.argv.includes('-w'),
+  MODE_TYPES = process.argv.includes('--config-types'),
+  EXTERNAL = ['react', 'react-dom', 'media-icons', 'media-captions', 'hls.js', /@radix-ui/],
+  NPM = [define({ dev: true }), define({ dev: false })];
+
+// Styles.
+copyStyles();
+copyTailwind();
+
+// Sandbox
+let isRunningSandbox = false;
+function launchSandbox() {
+  if (isRunningSandbox) return;
+  execa('pnpm', ['run', 'sandbox'], { stdio: 'inherit' });
+  isRunningSandbox = true;
+}
 
 export default defineConfig(
   MODE_WATCH ? [defineTypes(), ...NPM] : MODE_TYPES ? [defineTypes()] : NPM,
@@ -39,7 +60,9 @@ function defineTypes() {
       {
         name: 'cleanup',
         closeBundle() {
-          fs.rmSync('types', { recursive: true });
+          if (!MODE_WATCH) {
+            fs.rmSync('types', { recursive: true });
+          }
         },
       },
     ],
@@ -144,6 +167,23 @@ function define({ dev }) {
           }
         },
       },
+      dev && {
+        name: 'sandbox',
+        async closeBundle() {
+          await launchSandbox();
+        },
+      },
     ],
   };
+}
+
+function copyStyles() {
+  fs.copySync(VIDSTACK_PKG_PLAYER_STYLES_DIR, STYLES_DIR);
+}
+
+function copyTailwind() {
+  const tailwindFilePath = path.resolve(VIDSTACK_PKG_DIR, 'tailwind.cjs'),
+    tailwindDTSFilePath = path.resolve(VIDSTACK_PKG_DIR, 'tailwind.d.cts');
+  fs.copyFileSync(tailwindFilePath, path.resolve(ROOT_DIR, 'tailwind.cjs'));
+  fs.copyFileSync(tailwindDTSFilePath, path.resolve(ROOT_DIR, 'tailwind.d.cts'));
 }
